add config for arch linux

+++ .bash_profile
@@ -9,20 +9,28 @@ if [ -d "$HOME/bin" ] ; then
   PATH="$HOME/bin:$PATH"
 fi
 
-export EDITOR=vim
+if [ -d "$HOME/.local/bin" ] ; then 
+    PATH="$HOME/.local/bin:$PATH"
+fi
 
-export PATH="/usr/local/mysql/bin:$PATH"
-
+if [ -d "/usr/local/mysql/bin" ] ; then
+    export PATH="/usr/local/mysql/bin:$PATH"
+fi
 # Setting PATH for Python 3.4
 # The orginal version is saved in .bash_profile.pysave
 #PATH="/Library/Frameworks/Python.framework/Versions/3.4/bin:${PATH}"
-export PATH="/usr/local/bin:$PATH"
+#export PATH="/usr/local/bin:$PATH"
 
-alias tmux='tmux -2'
-alias showFiles='defaults write com.apple.finder AppleShowAllFiles YES; killall Finder /System/Library/CoreServices/Finder.app'
-alias hideFiles='defaults write com.apple.finder AppleShowAllFiles NO; killall Finder /System/Library/CoreServices/Finder.app'
+
+if [[ $OSTYPE == "darwin"* ]] ; then
+    alias showFiles='defaults write com.apple.finder AppleShowAllFiles YES; killall Finder /System/Library/CoreServices/Finder.app'
+    alias hideFiles='defaults write com.apple.finder AppleShowAllFiles NO; killall Finder /System/Library/CoreServices/Finder.app'
+    alias vim='mvim'
+fi
+
+export EDITOR=vim
 export CLICOLOR=1
-alias vim='mvim'
-alias goowork='~/Dev/djangobox/'
 
+alias gotowork='~/Dev/djangobox/'
+#alias tmux='tmux -2'
